backend/requirements.txt
Django>=3.0,<4.0
requests==2.28.1
djangorestframework==3.14.0
django-cors-headers==3.13.0